Replace uncaught handler in configuration interface
Replace the uncaught handler property in the configuration object
'onUncaughtException' with 'reportUncaughtExceptions' which is a
boolean flag used determine whether or not the uncaught exception
handler should be used. If true the flag will cause the uncaught
exception handler to be called on an uncaught exception and then
attempt to exit the application after trying to send the
relevant error to the API. This is the default behavior. If
false then the lib will not attach to the uncaught exception
event and not log these errors with the API.